vision-ms/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-ms"
version = "0.1.0"
description = "Multi-camera RTSP ingest service publishing frames to RabbitMQ."
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "motor==3.5.1",
    "pymongo==4.5.0",
    "numpy==1.26.4",
    "opencv-python-headless==4.9.0.80",
    "insightface==0.7.3",
    "onnxruntime-silicon==1.16.0",
    "pika==1.3.2",
    "pyyaml==6.0.1",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.6",
    "certifi==2025.11.12"
]

[project.scripts]
vision-ms = "src.main:run"
